requirements.txt
Django
psycopg2
django-debug-toolbar
django-devserver
django-donottrack
gunicorn
httplib2
python3-memcached
python3-openid
selenium
simplejson
pytz
requests
django-taggit
geocoder
icalendar
django_ical
vobject
markdown
